mission-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
hazelcast-python-client==5.4.0
python-jose[cryptography]==3.3.0
pydantic==2.9.2
pydantic-settings==2.6.1
